Fix open issues 45, 46, 50, and 51
