+++ docs/assets/dashboard-verification.png
